Day15 - From<Hash> -> usize
